quiz-platform/user-service/requirements.txt
flask==3.1.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
gunicorn==25.1.0
PyJWT==2.11.0
Werkzeug==3.1.6